[expo-dev-menu] Rearrange js code of `DevMenuItems`
