+++ packages/expo-dev-menu/app/views/DevMenuView.tsx
@@ -2,11 +2,11 @@ import React from 'react';
 import { StyleSheet, PixelRatio, View } from 'react-native';
 
 import DevMenuContext, { Context } from '../DevMenuContext';
-import { StyledText } from '../components/Text';
-import Colors from '../constants/Colors';
-import DevMenuItemsList from './DevMenuItemsList';
-import DevMenuAppInfo from './DevMenuAppInfo';
 import { DevMenuAppInfoType, DevMenuItemAnyType } from '../DevMenuInternal';
+import ListFooter from '../components/ListFooter';
+import MainOptions from '../components/MainOptions';
+import DevMenuGroup from '../components/items/DevMenuGroup';
+import DevMenuAppInfo from './DevMenuAppInfo';
 
 type Props = {
   appInfo: DevMenuAppInfoType;
@@ -26,7 +26,13 @@ class DevMenuView extends React.PureComponent<Props, undefined> {
   };
 
   renderItems() {
-    return <DevMenuItemsList items={this.context.devMenuItems} />;
+    return (
+      <View>
+        <DevMenuGroup items={this.context.devMenuItems} />
+        <MainOptions />
+        <ListFooter label="This development menu will not be present in any release builds of this project." />
+      </View>
+    );
   }
 
   renderContent() {
@@ -76,12 +82,6 @@ const styles = StyleSheet.create({
   },
   itemsContainer: {
     marginTop: 7,
-  },
-  closeButton: {
-    position: 'absolute',
-    right: 12,
-    top: 12,
-    zIndex: 3, // should be higher than zIndex of onboarding container
   },
 });
